+++ src/main/java/com/book/backend/domain/oidc/Provider.java
@@ -1,4 +1,4 @@
-package com.book.backend.util.oidc;
+package com.book.backend.domain.oidc;
 
 public enum Provider {
     APPLE, KAKAO
