+++ src/main/java/com/book/backend/domain/oidc/record/OidcPublicKeyList.java
@@ -1,4 +1,4 @@
-package com.book.backend.util.oidc.record;
+package com.book.backend.domain.oidc.record;
 
 import com.book.backend.exception.CustomException;
 import com.book.backend.exception.ErrorCode;
